+++ pymccorrelation/pymccorrelation.py
@@ -25,7 +25,7 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """
 
-__version__ = '0.2.2'
+__version__ = '0.2.3'
 
 import numpy as _np
 import scipy.stats as _st