+++ CHANGELOG.md
@@ -4,9 +4,15 @@
 
 ### 0.2.3 (in preparation)
 
+#### Enhancements
+
+- Check x/y limit arrays to ensure they contain sensible values.
+- Update assertions on array lengths to provide more helpful debugging information.
+
+#### Bugfixes
+
 - fix description of `return_dist` behavior in doctrsing for `pymccorrelation`
 - Amend changelog to fix release date for v0.2.2.
-- Check x/y limit arrays to ensure they contain sensible values.
 
 ### 0.2.2 (2020 July 08)