+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 - fix description of `return_dist` behavior in doctrsing for `pymccorrelation`
 - Amend changelog to fix release date for v0.2.2.
+- Fix missing pass of x/y limits through `pymckendall()` wrapper
 
 ### 0.2.2 (2020 July 08)